separate out sub components
